+++ src/libs/components/Upload/index.tsx
@@ -1,4 +1,9 @@
 /**
+ * Copyright (c) We-Mak.
+ *
+ * This source code is licensed under the MIT license found in the
+ * LICENSE file in the root directory of this source tree.
+ * **************
  * Upload
  * Simple upload file
  */